feat: added i3bar

+++ nvim.sh
@@ -3,22 +3,25 @@
 nvim_path = $HOME/gabriel/neovim
 
 echo "Removing old neovim..."
-sudo rm /usr/local/bin/nvim
-sudo rm -r /usr/local/share/nvim/
+if [ -z $(which nvim) ]
+then
+    sudo rm /usr/local/bin/nvim
+    sudo rm -r /usr/local/share/nvim/
+fi
 if [ -d $nvim_path ] 
 then
     sudo rm -rf $nvim_path
 fi
 
 echo "Installing dependencies"
-pacman --noconfirm --needed -S base-devel cmake unzip ninja tree-sitter curl >/dev/null 2>&1
+sudo pacman --noconfirm --needed -S base-devel cmake unzip ninja tree-sitter curl
 echo "Cloning neovim repo..."
-git clone git@github.com:neovim/neovim.git >/dev/null 2>&1
+git clone git@github.com:neovim/neovim.git
 
 echo "Installing Neovim, this will take a while..."
 cd $HOME/neovim
-make CMAKE_BUILD_TYPE=Release >/dev/null 2>&1
-sudo make install >/dev/null 2>&1
+make CMAKE_BUILD_TYPE=Release
+sudo make install
 cd $HOME
 
 echo "Neovim is installed"
